Refactor ApiFormattedEvent
Removes the ‘.build_from_response’ class method and changes ‘.new’ to
accept the response directly. The ‘attributes’ accessor now builds the
hash in real-time, rather than assigning it to a variable upfront.

This means we will be able to provide other functionality in the 
instance using parts of the response that aren’t present in ‘attributes’.