Check if the calendar has declined the event
In the case where a meeting room has been booked by multiple people at
the same time, Google Calendar will decline all but one of the events.
However, all are included in the API response.

This adds a check to ApiFormattedEvent to detect if an event has been
declined, by checking the ‘attendees’ array. The outcome is exposed in
the ‘#declined?’ method.