Add a Gap model
A Gap is a period of time between two events in a meeting room. It is 
not persisted in the database, but a future change will add the ability 
for an event list to be built which includes gaps at relevant places.

Gaps are value objects designed to behave a bit like events, so we can 
display them in a similar way to users.