Build a list of events with gaps inbetween
This adds a new presenter for the Room model, which implements the 
decorator pattern so we can add additional view-specific methods to a 
room.

This adds a new method '#events_with_gaps', which will return a list of
events with Gap instances inserted between events which do not directly
follow-on from each other. A Gap is also inserted at the start of the 
list if there is no current event.